feat/mainpageを消して、taskアプリケーションを作成
