--- project/code/mainpage/models.py
@@ -1,55 +1,0 @@
-from django.conf import settings
-from django.db import models
-
-class Project(models.Model):
-    name = models.CharField(max_length=255)
-    description = models.TextField()
-    owner = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name='owned_projects')
-
-    def __str__(self):
-        return self.name
-
-class ProjectMember(models.Model):
-    project = models.ForeignKey(Project, on_delete=models.CASCADE, related_name='members')
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name='projects')
-
-    def __str__(self):
-        return f'{self.user.username} in {self.project.name}'
-
-class ChatMessage(models.Model):
-    project = models.ForeignKey(Project, on_delete=models.CASCADE, related_name='messages')
-    sender = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    content = models.TextField()
-    timestamp = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return f'{self.sender.username}: {self.content[:20]}'
-
-class File(models.Model):
-    project = models.ForeignKey(Project, on_delete=models.CASCADE, related_name='files')
-    name = models.CharField(max_length=255)
-    path = models.CharField(max_length=255)
-    uploader = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    upload_timestamp = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return self.name
-
-class Task(models.Model):
-    project = models.ForeignKey(Project, on_delete=models.CASCADE, related_name='tasks')
-    name = models.CharField(max_length=255)
-    description = models.TextField()
-    assignee = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name='tasks')
-    due_date = models.DateTimeField()
-    progress = models.CharField(max_length=50)
-    tags = models.ManyToManyField('Tag', related_name='tasks')
-
-    def __str__(self):
-        return self.name
-
-class Tag(models.Model):
-    name = models.CharField(max_length=50)
-    color = models.CharField(max_length=7)  # Color in hex format, e.g., #RRGGBB
-
-    def __str__(self):
-        return self.name
