+++ project/code/task/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from task.models import Task
+
+# Register your models here.
+admin.site.register(Task)
