+++ project/code/task/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class MainpageConfig(AppConfig):
+class TaskConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'mainpage'
+    name = 'task'
