+++ project/code/yumeji/settings.py
@@ -38,7 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'mainpage.apps.MainpageConfig',
+    'task.apps.TaskConfig',
 ]
 
 MIDDLEWARE = [
